ENH: standard clean routine
